+++ apavedialog.tcl
@@ -42,7 +42,6 @@ package require Tk
 source [file join [file dirname [info script]] apave.tcl]
 
 namespace eval apave {
-  source [file join [file dirname [info script]] apaveimg.tcl]
 }
 
 
@@ -60,9 +59,6 @@ oo::class create apave::APaveDialog {
     set _pdg(ns) [namespace current]::
     # namespace in object namespace for safety of its 'most important' data
     namespace eval ${_pdg(ns)}PD {}
-    foreach icon {err info warn ques} {
-      image create photo ${_pdg(ns)}PD::img$icon -data [set apave::img_$icon]
-    }
     if {[llength [self next]]} { next {*}$args }
   }
 
@@ -485,7 +481,7 @@ oo::class create apave::APaveDialog {
     # layout: add the icon
     if {$icon ni {"" "-"}} {
       set widlist [list [list labBimg - - 99 1 \
-      "-st n -pady 7" "-image ${_pdg(ns)}PD::img$icon"]]
+      "-st n -pady 7" "-image [apave::iconImage $icon]"]]
       set prevl labBimg
     } else {
       set widlist [list [list labimg - - 99 1]]
@@ -578,14 +574,14 @@ oo::class create apave::APaveDialog {
       if {$readonly || $hidefind || $chmsg ne ""} {
         append binds "
           menu \$pop
-           \$pop add command -accelerator Ctrl+C -label \"Copy\" \\
+           \$pop add command [my IconA copy] -accelerator Ctrl+C -label \"Copy\" \\
             -command \"event generate $wt <<Copy>>\""
         if {$hidefind || $chmsg ne ""} {
           append binds "
             \$pop configure -tearoff 0
             \$pop add separator
-            \$pop add command -accelerator Ctrl+A -label \"Select All\" \\
-            -command \"$wt tag add sel 1.0 end\""
+            \$pop add command [my IconA none] -accelerator Ctrl+A \\
+            -label \"Select All\" -command \"$wt tag add sel 1.0 end\""
         }
       }
     }
@@ -594,6 +590,7 @@ oo::class create apave::APaveDialog {
         if {![info exists ${_pdg(ns)}PD::fnd]} {
           set ${_pdg(ns)}PD::fnd ""
         }
+        set noIMG "[my IconA none]"
         if {$hidefind} {
           lappend widlist [list h__ h_3 L 1 4 "-cw 1"]
         } else {
@@ -614,12 +611,12 @@ oo::class create apave::APaveDialog {
           if {!$hidefind} {
             append binds "
              \$pop add separator
-             \$pop add command -accelerator Ctrl+F -label \"Find first\" \\
-              -command \"[self] InitFindInText; focus \[[self] Entfind\]\"
-             \$pop add command -accelerator F3 -label \"Find next\" \\
+             \$pop add command [my IconA find] -accelerator Ctrl+F -label \\
+             \"Find first\" -command \"[self] InitFindInText; focus \[[self] Entfind\]\"
+             \$pop add command $noIMG -accelerator F3 -label \"Find next\" \\
               -command \"[self] FindInText 1\"
              \$pop add separator
-             \$pop add command -accelerator Esc -label \"Exit\" \\
+             \$pop add command [my IconA exit] -accelerator Esc -label \"Exit\" \\
               -command \"\[[self] Pdg defb1\] invoke\"
             "
           }
@@ -633,29 +630,29 @@ oo::class create apave::APaveDialog {
             bind $wt <Alt-Up>    {[self] LinesMove -1}
             bind $wt <Alt-Down>  {[self] LinesMove +1}
             menu \$pop
-             \$pop add command -accelerator Ctrl+X -label \"Cut\" \\
+             \$pop add command [my IconA cut] -accelerator Ctrl+X -label \"Cut\" \\
               -command \"event generate $wt <<Cut>>\"
-             \$pop add command -accelerator Ctrl+C -label \"Copy\" \\
+             \$pop add command [my IconA copy] -accelerator Ctrl+C -label \"Copy\" \\
               -command \"event generate $wt <<Copy>>\"
-             \$pop add command -accelerator Ctrl+V -label \"Paste\" \\
+             \$pop add command [my IconA paste] -accelerator Ctrl+V -label \"Paste\" \\
               -command \"event generate $wt <<Paste>>\"
              \$pop add separator
-             \$pop add command -accelerator Ctrl+D -label \"Double line(s)\" \\
+             \$pop add command [my IconA double] -accelerator Ctrl+D -label \"Double selection\" \\
               -command \"[self] DoubleText 0\"
-             \$pop add command -accelerator Ctrl+Y -label \"Delete a line\" \\
+             \$pop add command [my IconA delete] -accelerator Ctrl+Y -label \"Delete line\" \\
               -command \"[self] DeleteLine 0\"
-             \$pop add command -accelerator Alt+Up -label \"Line(s) up\" \\
+             \$pop add command [my IconA up] -accelerator Alt+Up -label \"Line(s) up\" \\
               -command \"[self] LinesMove -1 0\"
-             \$pop add command -accelerator Alt+Down -label \"Line(s) down\" \\
+             \$pop add command [my IconA down] -accelerator Alt+Down -label \"Line(s) down\" \\
               -command \"[self] LinesMove +1 0\"
              \$pop add separator
-             \$pop add command -accelerator Ctrl+F -label \"Find first\" \\
+             \$pop add command [my IconA find] -accelerator Ctrl+F -label \"Find first\" \\
               -command \"[self] InitFindInText; focus \[[self] Entfind\]\"
-             \$pop add command -accelerator F3 -label \"Find next\" \\
+             \$pop add command $noIMG -accelerator F3 -label \"Find next\" \\
               -command \"[self] FindInText 1\"
              \$pop add separator
-             \$pop add command -accelerator Ctrl+W -label \"Save and exit\" \\
-              -command \"\[[self] Pdg defb1\] invoke\"
+             \$pop add command [my IconA SaveFile] -accelerator Ctrl+W \\
+             -label \"Save and exit\" -command \"\[[self] Pdg defb1\] invoke\"
             "
           oo::objdefine [self] export DoubleText DeleteLine LinesMove
         }
@@ -739,9 +736,13 @@ oo::class create apave::APaveDialog {
     oo::objdefine [self] unexport FindInText InitFindInText DoubleText DeleteLine Pdg
     set pdgeometry [winfo geometry $_pdg(win).dia]
     # the dialog's result is defined by "pave res" + checkbox's value
-    set res [my res $_pdg(win).dia]
-    if {$res && [set ${_pdg(ns)}PD::ch]} {
-      incr res 10
+    set res [set result [my res $_pdg(win).dia]]
+    set chv [set ${_pdg(ns)}PD::ch]
+    if { [string is integer $res] } {
+      if {$res && $chv} { incr result 10 }
+    } else {
+      set res [expr {$result ne "" ? 1 : 0}]
+      if {$res && $chv} { append result 10 }
     }
     if {$textmode && !$readonly} {
       set focusnow [my TexM]
@@ -773,9 +774,9 @@ oo::class create apave::APaveDialog {
       lassign [split $pdgeometry x+] w h x y
       if {abs($x-$gx)<30} {set x $gx}
       if {abs($y-$gy)<30} {set y $gy}
-      return [list $res ${w}x${h}+${x}+${y} $textcont [string trim $inopts]]
+      return [list $result ${w}x${h}+${x}+${y} $textcont [string trim $inopts]]
     }
-    return "$res$textcont$inopts"
+    return "$result$textcont$inopts"
 
   }