+++ apaveinput.tcl
@@ -34,7 +34,7 @@
 
 package require Tk
 
-package provide apave 2.5
+package provide apave 2.6
 
 source [file join [file dirname [info script]] apavedialog.tcl]